1) Returned AddressID and BookingID to int instead of long since the sql serial type generates int. 2) Completed SQL Inserts 3) Added IEnumerable<Customer> Owners to the CreditCard model 4) Added airlineID to ParseFlights (previously missed) 5) Added state to ParseAddress (previously missed) 6) Added iataID to GetCustomers (previously missed) 7) Fixed iata_id in GetAirport (were previously using iataid when the database uses iata_id)

+++ application/FlightBooking/FlightBooking/Models/Address.cs
@@ -14,9 +14,10 @@ namespace FlightBooking.Models
         [DisplayName("Zip Code")]
         public string ZipCode { get; set; }
         public string Country { get; set; }
-        public long AddressID { get; set; }
+        public int AddressID { get; set; }
+        // TODO: Add IEnumberable<Customer> for the customers the address is home to?
 
-        public Address(int streetNumber, string streetName, string city, string state, string zipCode, string country, long addressID)
+        public Address(int streetNumber, string streetName, string city, string state, string zipCode, string country, int addressID)
         {
             StreetNumber = streetNumber;
             StreetName = streetName;
@@ -27,7 +28,7 @@ namespace FlightBooking.Models
             AddressID = addressID;
         }
         
-        public Address(int streetNumber, string streetName, string city, string country, long addressID)
+        public Address(int streetNumber, string streetName, string city, string country, int addressID)
         {
             StreetNumber = streetNumber;
             StreetName = streetName;
